ImageSearchProject/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
google-cloud-storage
google-cloud-vision
firebase-admin